fix: vector store + in memory vector store example
