+++ example/src/wikipedia_vector.ts
@@ -1,0 +1,75 @@
+import { createVectorStoreTools, EmbeddingResult } from '@fabrice-ai/tools/vector'
+import { agent } from 'fabrice-ai/agent'
+import { teamwork } from 'fabrice-ai/teamwork'
+import { logger } from 'fabrice-ai/telemetry'
+import { solution, workflow } from 'fabrice-ai/workflow'
+
+import { lookupWikipedia } from './tools/wikipedia.js'
+
+const createInMemoryVectorStore = () => {
+  /**
+   * In-memory implementation of the VectorStore interface using functions.
+   */
+  const store = new Map<string, EmbeddingResult>()
+
+  const set = async (id: string, value: EmbeddingResult): Promise<void> => {
+    store.set(id, value)
+  }
+
+  const entries = async (): Promise<[string, EmbeddingResult][]> => {
+    return Array.from(store.entries())
+  }
+
+  return {
+    set,
+    entries,
+  }
+}
+const { saveDocumentInVectorStore, searchInVectorStore } = createVectorStoreTools(
+  createInMemoryVectorStore()
+)
+
+const wikipediaIndexer = agent({
+  role: 'Wikipedia Indexer',
+  description: `
+    You are skilled at reading and understanding the context of Wikipedia articles.
+    You split Wikipedia articles by each sentence to make it easy for other team members to search exact sentences in vector store.
+  `,
+  tools: {
+    lookupWikipedia,
+    saveDocumentInVectorStore,
+  },
+})
+
+const reportCompiler = agent({
+  role: 'Report Compiler',
+  description: `
+    You are skilled at compiling information from various sources into a coherent report.
+    You have access to Vector database with indexed sentences of Wikipedia articles.
+    You are making use of it to find relevant information.
+  `,
+  tools: {
+    searchInVectorStore,
+  },
+})
+
+const wikipediaResearch = workflow({
+  members: [wikipediaIndexer, reportCompiler],
+  description: `
+    Find information about John III Sobieski.
+    Index the data into vector database. One sentence is one document saved in Vector store.
+    List exact some example sentences related to:
+     - Battle of Vienna.
+     - John III Youth,
+     - John III later years and death.
+  `,
+  output: `
+    Report with:
+     - bullet points - list of 2 sentences per each topic from vector store.
+  `,
+  snapshot: logger,
+})
+
+const result = await teamwork(wikipediaResearch)
+
+console.log(solution(result))
